removed extraneous reference to stm32l4/nvic.h as the common header picks it up automatically
Added #ifndef around the ESOS BITx definitions in all_generic.h.  That did NOT fix the warnings as the ESOS headers is read before the libOpenCM3 header that contains the same defintion.  Might have to redefine the ESOS definitions to be something like "_esosBITx".  I don't want to do that as that is not very intuitive, especially is the port does not have a BITx defined somewhere.